nav bar with all items, breaking apart project tree and panel layout success
