+++ frontend/src/layout/panel-layout/panelLayoutLogic.tsx
@@ -1,7 +1,10 @@
 import { kea } from 'kea'
+import { LemonTreeRef } from 'lib/lemon-ui/LemonTree/LemonTree'
 
 import { navigation3000Logic } from '../navigation-3000/navigationLogic'
 import type { panelLayoutLogicType } from './panelLayoutLogicType'
+
+export type PanelLayoutNavItem = 'project' | 'activity' | 'persons'
 
 export const panelLayoutLogic = kea<panelLayoutLogicType>({
     path: ['layout', 'panel-layout', 'panelLayoutLogic'],
@@ -14,7 +17,10 @@ export const panelLayoutLogic = kea<panelLayoutLogicType>({
         toggleLayoutPanelPinned: (pinned: boolean) => ({ pinned }),
         // TODO: This is a temporary action to set the active navbar item
         // We should remove this once we have a proper way to handle the navbar item
-        setActiveLayoutNavBarItem: (item: 'project' | 'activity') => ({ item }),
+        setActiveLayoutNavBarItem: (item: PanelLayoutNavItem) => ({ item }),
+        setSearchTerm: (searchTerm: string) => ({ searchTerm }),
+        clearSearch: true,
+        setPanelTreeRef: (ref: React.RefObject<LemonTreeRef>) => ({ ref }),
     },
     reducers: {
         isLayoutNavbarVisibleForDesktop: [
@@ -38,7 +44,26 @@ export const panelLayoutLogic = kea<panelLayoutLogicType>({
                 toggleLayoutPanelPinned: (_, { pinned }) => pinned || _,
             },
         ],
-        isLayoutPanelPinned: [false, { toggleLayoutPanelPinned: (_, { pinned }) => pinned }],
-        activeLayoutNavBarItem: ['project', { setActiveLayoutNavBarItem: (_, { item }) => item }],
+        isLayoutPanelPinned: [
+            false,
+            { persist: true },
+            {
+                toggleLayoutPanelPinned: (_, { pinned }) => pinned,
+            },
+        ],
+        activeLayoutNavBarItem: ['', { setActiveLayoutNavBarItem: (_, { item }) => item }],
+        searchTerm: [
+            '',
+            {
+                setSearchTerm: (_, { searchTerm }) => searchTerm,
+                clearSearch: () => '',
+            },
+        ],
+        panelTreeRef: [
+            null,
+            {
+                setPanelTreeRef: (_: unknown, { ref }: { ref: React.RefObject<LemonTreeRef> }) => ref as any,
+            },
+        ] as any,
     },
 })
